requirements.txt
aiofiles==0.8.0
aiogoogle==3.2.1
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.49.0
frozenlist==1.3.0
google-auth==2.6.0
h11==0.13.0
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tonyg-rfc3339==0.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.6
yarl==1.7.2
